functions/anomaly_detect/requirements.txt
functions-framework==3.4.0
google-cloud-bigquery==3.14.0
google-generativeai==0.3.2
pandas==2.1.4
PyYAML==6.0.1
